Add files via upload
These are the files of Project of class 43 Treasure Hunt. There are 14 files and 1 Js folder this project. In this class, we have learnt about the games elements, making games fun and engaging the players. In this game we have designed a den in which the treasure is there and the player have to answer the 3 questions and after answering all questions correctly, the treasure will be given to the player.